// frontend/src/app/api/model/external-id-response.ts
/**
 * CloudPulse AI API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExternalIdResponse { 
    externalId: string;
    cloudFormationTemplateUrl: string;
}